PR suggestion - move the insight logic to collection-header
