+++ packages/compass-collection/src/components/collection-header/collection-header.tsx
@@ -13,6 +13,7 @@ import type { Signal } from '@mongodb-js/compass-components';
 import type { Document } from 'mongodb';
 import React, { Component } from 'react';
 import toNS from 'mongodb-ns';
+import { withPreferences } from 'compass-preferences-model';
 
 import CollectionHeaderActions from '../collection-header-actions';
 import ReadOnlyBadge from './read-only-badge';
@@ -108,12 +109,14 @@ const insightsContainerStyles = css({
 
 type CollectionHeaderProps = {
   darkMode?: boolean;
+  showInsights: boolean;
   globalAppRegistry: AppRegistry;
   namespace: string;
   isReadonly: boolean;
   isTimeSeries: boolean;
   isClustered: boolean;
   isFLE: boolean;
+  isAtlas: boolean;
   selectOrCreateTab: (options: any) => any;
   sourceName?: string;
   sourceReadonly?: boolean;
@@ -121,7 +124,46 @@ type CollectionHeaderProps = {
   editViewName?: string;
   pipeline: Document[];
   stats: CollectionStatsObject;
-  insights: Signal[];
+};
+
+const ATLAS_SEARCH_LP_LINK = 'https://www.mongodb.com/cloud/atlas/lp/search-1';
+
+const getInsightsForPipeline = (pipeline: Document[], isAtlas: boolean) => {
+  const insights: Record<string, Signal> = {};
+  for (const stage of pipeline) {
+    if ('$match' in stage) {
+      const stringifiedStageValue = JSON.stringify(stage);
+      if (
+        stringifiedStageValue.includes('$regex') ||
+        stringifiedStageValue.includes('$text')
+      ) {
+        const signal = isAtlas
+          ? {
+              id: 'atlas-text-regex-usage-in-view',
+              title: 'Alternate text search options available',
+              description:
+                "In many cases, Atlas Search is MongoDB's most efficient full text search option. Convert your view’s query to $search for a wider range of functionality.",
+              learnMoreLink:
+                'https://www.mongodb.com/docs/atlas/atlas-search/best-practices/',
+              primaryActionButtonLink: ATLAS_SEARCH_LP_LINK,
+              primaryActionButtonLabel: 'Try Atlas Search',
+            }
+          : {
+              id: 'non-atlas-text-regex-usage-in-view',
+              title: 'Alternate text search options available',
+              description:
+                "In many cases, Atlas Search is MongoDB's most efficient full text search option. Connect with Atlas to explore the power of Atlas Search.",
+              learnMoreLink:
+                'https://www.mongodb.com/docs/atlas/atlas-search/best-practices/',
+              primaryActionButtonLink: ATLAS_SEARCH_LP_LINK,
+              primaryActionButtonLabel: 'Try Atlas Search',
+            };
+
+        insights[signal.id] = signal;
+      }
+    }
+  }
+  return Object.values(insights);
 };
 
 class CollectionHeader extends Component<CollectionHeaderProps> {
@@ -170,7 +212,10 @@ class CollectionHeader extends Component<CollectionHeaderProps> {
     const ns = toNS(this.props.namespace);
     const database = ns.database;
     const collection = ns.collection;
-
+    const insights =
+      this.props.showInsights && this.props.pipeline?.length
+        ? getInsightsForPipeline(this.props.pipeline, this.props.isAtlas)
+        : [];
     return (
       <div
         className={cx(
@@ -226,9 +271,9 @@ class CollectionHeader extends Component<CollectionHeaderProps> {
           {this.props.isClustered && <ClusteredBadge />}
           {this.props.isFLE && <FLEBadge />}
           {this.props.isReadonly && this.props.sourceName && <ViewBadge />}
-          {!!this.props.insights.length && (
+          {!!insights.length && (
             <div className={insightsContainerStyles}>
-              <SignalPopover signals={this.props.insights} />
+              <SignalPopover signals={insights} />
             </div>
           )}
           <CollectionHeaderActions
@@ -250,4 +295,8 @@ class CollectionHeader extends Component<CollectionHeaderProps> {
   }
 }
 
-export default withDarkMode(CollectionHeader);
+export default withPreferences(
+  withDarkMode(CollectionHeader),
+  ['showInsights'],
+  React
+);
